Added an AvroTopicConfig
